# Speeding up log scrubbing regex matching